+++ app/components/keyboard/modal/VariantForm.vue
@@ -234,8 +234,6 @@ const variant = ref({
   img_front: '',
   img_back: '',
   photo_credit: '',
-  mount_styles: null,
-  typing_angle: null,
   currency: 'USD',
   msrp_price: null,
   case_materials: [],
@@ -260,10 +258,6 @@ const schema = z.object({
   img_front: z.url().nullish().or(z.string().min(0).max(0)),
   img_back: z.url().nullish().or(z.string().min(0).max(0)),
   photo_credit: z.string().max(255).nullish().or(z.string().min(0).max(0)),
-  mount_styles: z
-    .array(z.enum(Constants.public.Enums.keyboard_mounting_style))
-    .nullish(),
-  typing_angle: z.coerce.number().min(0).max(30).nullish(),
   currency: z.enum(currencies).nullish().or(z.string().min(0).max(0)),
   msrp_price: z.coerce.number().min(0).nullish(),
   case_materials: z